pico: split multiplayer from usb_device
Replacing the multiplayer stubs in usb_host with cdc stubs
