+++ 32blit-pico/usb.hpp
@@ -7,7 +7,8 @@ void update_usb();
 
 void usb_debug(const char *message);
 
-// TODO: separate multiplayer from usb
-bool is_multiplayer_connected();
-void set_multiplayer_enabled(bool enabled);
-void send_multiplayer_message(const uint8_t *data, uint16_t len);
+bool usb_cdc_connected();
+uint16_t usb_cdc_read(uint8_t *data, uint16_t len);
+uint32_t usb_cdc_read_available();
+void usb_cdc_write(const uint8_t *data, uint16_t len);
+void usb_cdc_flush_write();
